# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Pi Wallet Auto-Forwarder API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Wallet {
  id: number;
  /** Friendly name for this wallet */
  label: string;
  /** @nullable */
  sourceAddress?: string | null;
  /** @nullable */
  destinationAddress?: string | null;
  isConfigured: boolean;
  hasSecretKey: boolean;
  pollIntervalSeconds: number;
  monitorRunning: boolean;
  /** @nullable */
  lastCheckedAt?: string | null;
  /** @nullable */
  lastError?: string | null;
  createdAt: string;
  /** Total Pi forwarded by this wallet */
  totalForwarded?: string;
  transferCount?: number;
}

export interface WalletInput {
  /** @minLength 1 */
  label: string;
  /** @minLength 1 */
  sourceAddress: string;
  /** @minLength 1 */
  destinationAddress: string;
  /** @minLength 1 */
  secretKey: string;
  /**
     * @minimum 10
     * @maximum 3600
     */
  pollIntervalSeconds?: number;
}

export interface WalletUpdate {
  /** @minLength 1 */
  label?: string;
  sourceAddress?: string;
  destinationAddress?: string;
  /** @nullable */
  secretKey?: string | null;
  /**
     * @minimum 10
     * @maximum 3600
     */
  pollIntervalSeconds?: number;
}

export interface MonitorSummary {
  totalWallets: number;
  runningWallets: number;
  totalPiForwarded: string;
  totalTransactions: number;
  successCount?: number;
  failedCount?: number;
}

export type TransferStatus = typeof TransferStatus[keyof typeof TransferStatus];


export const TransferStatus = {
  pending: 'pending',
  forwarded: 'forwarded',
  failed: 'failed',
} as const;

export interface Transfer {
  id: number;
  walletId: number;
  /** @nullable */
  walletLabel?: string | null;
  incomingTxHash: string;
  /** @nullable */
  outgoingTxHash?: string | null;
  amount: string;
  /** @nullable */
  fromAddress?: string | null;
  status: TransferStatus;
  /** @nullable */
  errorMessage?: string | null;
  createdAt: string;
  /** @nullable */
  forwardedAt?: string | null;
}

export interface TransferList {
  transfers: Transfer[];
  total: number;
}

export interface TransferStats {
  totalPiForwarded: string;
  totalTransactions: number;
  successCount: number;
  failedCount: number;
  pendingCount: number;
  recentTransfers?: Transfer[];
}

export type ListTransfersParams = {
/**
 * Filter by wallet ID
 */
walletId?: number;
limit?: number;
offset?: number;
};