# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Pi Wallet Auto-Forwarder API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: wallets
    description: Wallet management
  - name: transfers
    description: Transfer history
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /wallets:
    get:
      operationId: listWallets
      tags: [wallets]
      summary: List all monitored wallets
      responses:
        "200":
          description: List of wallets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Wallet"
    post:
      operationId: createWallet
      tags: [wallets]
      summary: Add a new wallet to monitor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletInput"
      responses:
        "201":
          description: Created wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
        "400":
          description: Validation error

  /wallets/{id}:
    get:
      operationId: getWallet
      tags: [wallets]
      summary: Get a wallet by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
        "404":
          description: Not found
    put:
      operationId: updateWallet
      tags: [wallets]
      summary: Update a wallet's configuration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletUpdate"
      responses:
        "200":
          description: Updated wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
        "400":
          description: Validation error
        "404":
          description: Not found
    delete:
      operationId: deleteWallet
      tags: [wallets]
      summary: Remove a wallet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  /wallets/{id}/start:
    post:
      operationId: startWalletMonitor
      tags: [wallets]
      summary: Start monitoring a wallet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Wallet with updated monitor state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
        "400":
          description: Cannot start (missing config)
        "404":
          description: Not found

  /wallets/{id}/stop:
    post:
      operationId: stopWalletMonitor
      tags: [wallets]
      summary: Stop monitoring a wallet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Wallet with updated monitor state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
        "404":
          description: Not found

  /monitor/summary:
    get:
      operationId: getMonitorSummary
      tags: [wallets]
      summary: Get overall monitoring summary across all wallets
      responses:
        "200":
          description: Summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonitorSummary"

  /transfers:
    get:
      operationId: listTransfers
      tags: [transfers]
      summary: List transfer records
      parameters:
        - name: walletId
          in: query
          schema:
            type: integer
          description: Filter by wallet ID
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of transfers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferList"

  /transfers/stats:
    get:
      operationId: getTransferStats
      tags: [transfers]
      summary: Aggregate transfer stats across all wallets
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferStats"

  /transfers/{id}:
    get:
      operationId: getTransfer
      tags: [transfers]
      summary: Get a single transfer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transfer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transfer"
        "404":
          description: Not found

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    Wallet:
      type: object
      required: [id, label, isConfigured, hasSecretKey, pollIntervalSeconds, monitorRunning, createdAt]
      properties:
        id:
          type: integer
        label:
          type: string
          description: Friendly name for this wallet
        sourceAddress:
          type: ["string", "null"]
        destinationAddress:
          type: ["string", "null"]
        isConfigured:
          type: boolean
        hasSecretKey:
          type: boolean
        pollIntervalSeconds:
          type: integer
        monitorRunning:
          type: boolean
        lastCheckedAt:
          type: ["string", "null"]
        lastError:
          type: ["string", "null"]
        createdAt:
          type: string
        totalForwarded:
          type: string
          description: Total Pi forwarded by this wallet
        transferCount:
          type: integer

    WalletInput:
      type: object
      required: [label, sourceAddress, destinationAddress, secretKey]
      properties:
        label:
          type: string
          minLength: 1
        sourceAddress:
          type: string
          minLength: 1
        destinationAddress:
          type: string
          minLength: 1
        secretKey:
          type: string
          minLength: 1
        pollIntervalSeconds:
          type: integer
          minimum: 10
          maximum: 3600

    WalletUpdate:
      type: object
      properties:
        label:
          type: string
          minLength: 1
        sourceAddress:
          type: string
        destinationAddress:
          type: string
        secretKey:
          type: ["string", "null"]
        pollIntervalSeconds:
          type: integer
          minimum: 10
          maximum: 3600

    MonitorSummary:
      type: object
      required: [totalWallets, runningWallets, totalPiForwarded, totalTransactions]
      properties:
        totalWallets:
          type: integer
        runningWallets:
          type: integer
        totalPiForwarded:
          type: string
        totalTransactions:
          type: integer
        successCount:
          type: integer
        failedCount:
          type: integer

    Transfer:
      type: object
      required: [id, walletId, incomingTxHash, amount, status, createdAt]
      properties:
        id:
          type: integer
        walletId:
          type: integer
        walletLabel:
          type: ["string", "null"]
        incomingTxHash:
          type: string
        outgoingTxHash:
          type: ["string", "null"]
        amount:
          type: string
        fromAddress:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, forwarded, failed]
        errorMessage:
          type: ["string", "null"]
        createdAt:
          type: string
        forwardedAt:
          type: ["string", "null"]

    TransferList:
      type: object
      required: [transfers, total]
      properties:
        transfers:
          type: array
          items:
            $ref: "#/components/schemas/Transfer"
        total:
          type: integer

    TransferStats:
      type: object
      required: [totalPiForwarded, totalTransactions, successCount, failedCount, pendingCount]
      properties:
        totalPiForwarded:
          type: string
        totalTransactions:
          type: integer
        successCount:
          type: integer
        failedCount:
          type: integer
        pendingCount:
          type: integer
        recentTransfers:
          type: array
          items:
            $ref: "#/components/schemas/Transfer"