# lib/api-zod/src/generated/types/monitorSummary.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Pi Wallet Auto-Forwarder API
 * OpenAPI spec version: 0.1.0
 */

export interface MonitorSummary {
  totalWallets: number;
  runningWallets: number;
  totalPiForwarded: string;
  totalTransactions: number;
  successCount?: number;
  failedCount?: number;
}