# lib/api-zod/src/generated/types/transfer.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Pi Wallet Auto-Forwarder API
 * OpenAPI spec version: 0.1.0
 */
import type { TransferStatus } from './transferStatus';

export interface Transfer {
  id: number;
  walletId: number;
  /** @nullable */
  walletLabel?: string | null;
  incomingTxHash: string;
  /** @nullable */
  outgoingTxHash?: string | null;
  amount: string;
  /** @nullable */
  fromAddress?: string | null;
  status: TransferStatus;
  /** @nullable */
  errorMessage?: string | null;
  createdAt: string;
  /** @nullable */
  forwardedAt?: string | null;
}