# lib/api-zod/src/generated/types/wallet.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Pi Wallet Auto-Forwarder API
 * OpenAPI spec version: 0.1.0
 */

export interface Wallet {
  id: number;
  /** Friendly name for this wallet */
  label: string;
  /** @nullable */
  sourceAddress?: string | null;
  /** @nullable */
  destinationAddress?: string | null;
  isConfigured: boolean;
  hasSecretKey: boolean;
  pollIntervalSeconds: number;
  monitorRunning: boolean;
  /** @nullable */
  lastCheckedAt?: string | null;
  /** @nullable */
  lastError?: string | null;
  createdAt: string;
  /** Total Pi forwarded by this wallet */
  totalForwarded?: string;
  transferCount?: number;
}