# lib/api-zod/src/generated/types/walletInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Pi Wallet Auto-Forwarder API
 * OpenAPI spec version: 0.1.0
 */

export interface WalletInput {
  /** @minLength 1 */
  label: string;
  /** @minLength 1 */
  sourceAddress: string;
  /** @minLength 1 */
  destinationAddress: string;
  /** @minLength 1 */
  secretKey: string;
  /**
     * @minimum 10
     * @maximum 3600
     */
  pollIntervalSeconds?: number;
}