# lib/api-zod/src/generated/types/walletUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Pi Wallet Auto-Forwarder API
 * OpenAPI spec version: 0.1.0
 */

export interface WalletUpdate {
  /** @minLength 1 */
  label?: string;
  sourceAddress?: string;
  destinationAddress?: string;
  /** @nullable */
  secretKey?: string | null;
  /**
     * @minimum 10
     * @maximum 3600
     */
  pollIntervalSeconds?: number;
}